% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobbs_douglas_mult.R
\name{cobb_douglas_mult}
\alias{cobb_douglas_mult}
\title{Solve for Y}
\usage{
cobb_douglas_mult(a, px, I)
}
\arguments{
\item{a}{a vector of alpha values}

\item{px}{a vector of prices}

\item{I}{the budget}
}
\value{
A dataframe
}
\description{
Solve for Y
}
\examples{
# 4 goods
I = 100
a = c(0.1, 0.1, 0.4, 0.4)
sum(a)
px = c(1,1,2,2)
cobb_douglas_mult(a = a, px = px, I = I)

# if you want to visualize 2 goods you can remove the extra-goods from the Budget
# let's focus on good x3 and x4
I_new = 100 - ((20*2) + (20*2))

# x3 and x4
I = I_new
a = c(0.4, 0.4) # keep the same alpha values
px = c(2,2) # keep the same prices

cobb_douglas_mult(a = a, px = px, I = I_new)
# we can plot a bivariate plot for x3 and x4
cobbs_douglas_utility(I = I_new, a = 0.4, b = 0.4, px = 2, py = 2)$optimal_bundle
}
