% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobbs_douglas_models.R
\name{cobbs_douglas_models}
\alias{cobbs_douglas_models}
\title{Compare Cobbs-Douglas Models}
\usage{
cobbs_douglas_models(..., ymax = 120, xmax = 120)
}
\arguments{
\item{...}{a list of models generated by the function cobbs_douglas_utility()}
}
\value{
A dataframe
}
\description{
Compare Cobbs-Douglas Models
}
\examples{
# Compare 4 models
model1 = cobbs_douglas_utility(I = 100, a = 0.25, b = 0.75, px = 1, py = 1)
model2 = cobbs_douglas_utility(I = 100, a = 0.25, b = 0.75, px = 1, py = 2)
model3 = cobbs_douglas_utility(I = 100, a = 0.25, b = 0.75, px = 2, py = 1)
model4_policy = cobbs_douglas_utility(I = 200, a = 0.25, b = 0.75, px = 2, py = 3)
cobbs_douglas_models(model1, model2, model3, model4_policy)
}
