% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobbs_douglas_utility.R
\name{cobbs_douglas_utility}
\alias{cobbs_douglas_utility}
\title{Calculate the optimal bundle for two goods (x and y) given a budget constraint.
\deqn{U(x,y) = x^a \cdot y^b}}
\usage{
cobbs_douglas_utility(I, a = 0.5, b = 0.5, px, py)
}
\arguments{
\item{I}{the budget constraint}

\item{a}{the alpha value}

\item{b}{the beta value}

\item{px}{the price of good x}

\item{py}{the price of good y}
}
\value{
A list.
}
\description{
Calculate the optimal bundle for two goods (x and y) given a budget constraint.
\deqn{U(x,y) = x^a \cdot y^b}
}
\examples{
cobbs_douglas_utility(I=100, a=0.5, b=0.5, px = 1, py = 2)
cobbs_douglas_utility(I=100, a=0.4, b=0.5, px = 1.2, py = 2)
}
