% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand_curve.R
\name{demand_curve}
\alias{demand_curve}
\title{Solve for Y}
\usage{
demand_curve(cb_objet, max_price = 50, y_max = 50, x_max = 50)
}
\arguments{
\item{U}{the Utility}

\item{x}{the quantity of x}

\item{a}{the alpha value}

\item{b}{the beta value}
}
\value{
A dataframe
}
\description{
Solve for Y
}
\examples{
a = 0.2; b = 0.5; px = 1; py = 3; I = 1000
cb_objet = cobbs_douglas_utility(I = I, a = a, b = a, px = px, py = py)
}
