% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobbs_douglas_utility.R
\name{f_cobbs_y}
\alias{f_cobbs_y}
\alias{f_cobbs}
\title{Solve for Y}
\usage{
f_cobbs(x = x, y = y, a = a, b = b)
}
\arguments{
\item{x}{the quantity of x}

\item{a}{the alpha value}

\item{b}{the beta value}

\item{U}{the Utility}
}
\value{
A dataframe
}
\description{
Solve for Y
}
\examples{
f_cobbs_y(x=0:10,U=5,a=0.5,b=0.5)
}
