% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivative.R
\name{f_derive}
\alias{f_derive}
\title{Solve for Y}
\usage{
f_derive(h = 0.01, f = f_cobbs, x, y, a, b)
}
\arguments{
\item{h}{the instantaneous rate of change}

\item{x}{the quantity of good x}

\item{y}{the quantity of good y}

\item{a}{the alpha value}

\item{b}{the beta value}
}
\value{
A dataframe
}
\description{
Solve for Y
}
\examples{
f_derive(h = 0.01, x = 10, y = 10, a = 0.5, b = 0.5) 
MUxy = - f_derive(h = 0.01, x = x, y = y, a = a, b = b) / f_derive(h = 0.01, x = y, y = x, a = b, b = a)
}
