% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_utility.R
\name{MU_cb}
\alias{MU_cb}
\title{Solve for Y}
\usage{
MU_cb(x, a)
}
\arguments{
\item{x}{the quantity of x}

\item{a}{the alpha value}
}
\value{
numeric
}
\description{
Solve for Y
}
\examples{
cb = cobbs_douglas_utility(I = 100, a = a, b = b, px = 1, py = 1)
a = 0.5
b = 0.5
x = cb$x
y = cb$y
plot(x, MU_cb(x, 0.5), type = 'l')
}
